+++ examples/exact.rs
@@ -1,8 +1,8 @@
 // Formats and parses an exact interval literal.
 
-use inari::{interval, Interval, IntervalError};
+use inari::{interval, IntervalError};
 
-fn main() -> Result<(), IntervalError<Interval>> {
+fn main() -> Result<(), IntervalError> {
     let x = interval!("[0.1]")?;
     let s = format!("{:x}", x);
     let y = interval!(&s, exact)?;